move NewGRPCClientConn to public package
